Time: 15 ms (36.12%), Space: 8.7 MB (36.60%) - LeetHub

+++ 0013-roman-to-integer/0013-roman-to-integer.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int romanToInt(string s) {
+    short romanToInt(string s) {
         unordered_map<char,short>map;
         map['I'] = 1;
         map['V'] = 5;
